steamshim: update file and header names to reflect the reversed relationship
