+++ source/server/CMakeLists.txt
@@ -3,7 +3,7 @@ project(${QFUSION_SERVER_NAME})
 include_directories(${MINIZ_INCLUDE_DIR} ${CURL_INCLUDE_DIR})
 
 if (BUILD_STEAMLIB)
-    set(STEAMSHIMCHILD_LIBRARY steamshim_child)
+    set(STEAMSHIMPARENT_LIBRARY steamshim_parent)
 endif()
 
 
@@ -91,7 +91,7 @@ endif()
 
 add_executable(${QFUSION_SERVER_NAME} ${SERVER_BINARY_TYPE} ${SERVER_HEADERS} ${SERVER_SOURCES} ${SERVER_PLATFORM_SOURCES})
 add_dependencies(${QFUSION_SERVER_NAME} angelwrap game)
-target_link_libraries(${QFUSION_SERVER_NAME} PRIVATE ${CURL_LIBRARY} ${SERVER_PLATFORM_LIBRARIES} ${STEAMSHIMCHILD_LIBRARY})
+target_link_libraries(${QFUSION_SERVER_NAME} PRIVATE ${CURL_LIBRARY} ${SERVER_PLATFORM_LIBRARIES} ${STEAMSHIMPARENT_LIBRARY})
 qf_set_output_dir(${QFUSION_SERVER_NAME} "")
 
 set_target_properties(${QFUSION_SERVER_NAME} PROPERTIES COMPILE_DEFINITIONS "DEDICATED_ONLY")
