+++ source/server/sv_steam.c
@@ -1,6 +1,6 @@
 #include "../qcommon/qcommon.h"
 #include "../qcommon/steam.h"
-#include "../steamshim/src/child/child.h"
+#include "../steamshim/src/parent/parent.h"
 #include <string.h>
 
 static void printEvent( const STEAMSHIM_Event *e )
