+++ source/tv_server/CMakeLists.txt
@@ -3,7 +3,7 @@ project(${QFUSION_TVSERVER_NAME})
 include_directories(${MINIZ_INCLUDE_DIR} ${CURL_INCLUDE_DIR})
 
 if (BUILD_STEAMLIB)
-set(STEAMSHIMCHILD_LIBRARY steamshim_child)
+    set(STEAMSHIMPARENT_LIBRARY steamshim_parent)
 endif()
 
 
@@ -86,7 +86,7 @@ else()
 endif()
 
 add_executable(${QFUSION_TVSERVER_NAME} ${TV_SERVER_BINARY_TYPE} ${TV_SERVER_HEADERS} ${TV_SERVER_SOURCES} ${TV_SERVER_PLATFORM_SOURCES})
-target_link_libraries(${QFUSION_TVSERVER_NAME} PRIVATE ${CURL_LIBRARY} ${TV_SERVER_PLATFORM_LIBRARIES} ${STEAMSHIMCHILD_LIBRARY})
+target_link_libraries(${QFUSION_TVSERVER_NAME} PRIVATE ${CURL_LIBRARY} ${TV_SERVER_PLATFORM_LIBRARIES} ${STEAMSHIMPARENT_LIBRARY})
 
 qf_set_output_dir(${QFUSION_TVSERVER_NAME} "")
